# code/backend/src/config/swagger.yaml
openapi: 3.0.0
info:
  title: API Reside
  version: 1.0.0
  description: API do Sistema de Gerenciamento de Condomínios Reside
servers:
  - url: https://reside-production.up.railway.app
    description: Servidor de Desenvolvimento Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Uma mensagem descritiva do erro.
          example: 'Token inválido ou expirado.'
...
    Message:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        receiverId:
          type: integer
          nullable: true
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /messages/residents/{senderId}/condominium-message:
    post:
      summary: Envia mensagem para todos os moradores do condomínio (broadcast)
      tags:
        - Messages
      parameters:
        - in: path
          name: senderId
          required: true
          schema:
            type: string
          description: ID do morador remetente
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                content:
                  type: string
                  description: Conteúdo da mensagem
                image:
                  type: string
                  format: binary
                  description: Imagem opcional (JPEG ou PNG, até 5MB)
              required:
                - content
      responses:
        '201':
          description: Mensagem enviada para todos os moradores do condomínio
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  messagesSent:
                    type: integer
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        '400':
          description: Conteúdo da mensagem ausente ou imagem inválida
        '500':
          description: Erro interno do servidor
  /messages/{messageId}/image:
    get:
      summary: Recupera a imagem anexada a uma mensagem
      tags:
        - Messages
      parameters:
        - in: path
          name: messageId
          required: true
          schema:
            type: string
          description: ID da mensagem
      responses:
        '200':
          description: Imagem da mensagem retornada com sucesso
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: Imagem não encontrada
        '500':
          description: Erro ao recuperar a imagem
  /messages/residents/{residentId}/messages:
    get:
      summary: Lista todas as mensagens recebidas por um morador
      tags:
        - Messages
      parameters:
        - in: path
          name: residentId
          required: true
          schema:
            type: string
          description: ID do morador
      responses:
        '200':
          description: Lista de mensagens recebidas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '500':
          description: Erro interno do servidor
  /messages/residents/{residentId}/messages/unread:
    get:
      summary: Lista todas as mensagens não lidas por um morador
      tags:
        - Messages
      parameters:
        - in: path
          name: residentId
          required: true
          schema:
            type: string
          description: ID do morador
      responses:
        '200':
          description: Lista de mensagens não lidas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '500':
          description: Erro interno do servidor
  /messages/{messageId}/read:
    put:
      summary: Marca uma mensagem como lida
      tags:
        - Messages
      parameters:
        - in: path
          name: messageId
          required: true
          schema:
            type: string
          description: ID da mensagem
      responses:
        '200':
          description: Mensagem marcada como lida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '500':
          description: Erro interno do servidor
...
  /notifications/test:
      post:
        tags:
          - Notificações
        summary: Enviar notificação de aluguel de vaga
        description: Dispara uma notificação push padrão “Vaga Alugada!” para o token FCM informado, usando o spotId para compor o texto.
        security:
          - bearerAuth: []
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationTestRequest'
              examples:
                teste:
                  summary: Exemplo de envio
                  value:
                    token: "fZWiH0XtRn2-EpaAcBvvsb:APA91bHmJIeKUwdARXuDTSu_sPTdZK8gZEwn9Z"
                    spotId: 401
        responses:
          '200':
            description: Notificação enviada com sucesso
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/NotificationTestResponse'
                examples:
                  sucesso:
                    summary: Resposta de sucesso
                    value:
                      success: true
                      message: "Notificação enviada!"
          '400':
            description: Parâmetros inválidos na requisição
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
                examples:
                  faltaParametros:
                    summary: Token ou spotId ausentes
                    value:
                      message: "token e spotId são obrigatórios"
          '401':
            description: Token de autenticação não fornecido ou inválido
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
                examples:
                  naoAutorizado:
                    summary: Falha de autenticação
                    value:
                      message: "Token de autenticação inválido"
          '500':
            description: Erro interno ao enviar a notificação
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
                examples:
                  erroServidor:
                    summary: Exceção no servidor
                    value:
                      message: "Erro disparando notificação"
...
  /notifications/parking-rented:
    post:
      tags:
        - Notificações
      summary: Enviar notificação de vaga alugada
      description: Envia uma notificação push específica para informar que uma vaga foi alugada
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ownerId:
                  type: integer
                  description: ID do proprietário da vaga que receberá a notificação
                spotId:
                  type: integer
                  description: ID da vaga que foi alugada
                renterName:
                  type: string
                  description: Nome completo da pessoa que alugou a vaga
              required:
                - ownerId
                - spotId
                - renterName
              example:
                ownerId: 456
                spotId: 123
                renterName: "João Silva"
      responses:
        '200':
          description: Notificação de vaga alugada enviada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Notificação enviada com sucesso!"
        '400':
          description: Parâmetros obrigatórios não fornecidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                parametrosObrigatorios:
                  summary: Parâmetros faltando
                  value:
                    error: "ownerId, spotId e renterName são obrigatórios"
        '500':
          description: Falha ao enviar notificação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                falhaNotificacao:
                  summary: Erro ao enviar notificação
                  value:
                    error: "Falha ao enviar notificação"